Add service to process inventory udpate
The commits has the code that will be used to udpate/create the
inventory data on our DB.

The commit also has the gem `deterministic`, which has a monadic handle
to our core functionalities. Using a monadic core helps us avoid the
necesity of add `try/catch` in all possible places.